move RelatedPostLayout to own component
